opengl fucking sucks
after much screaming and gnashing of teeth, i've finally gotten opengl to work more or less in tandem with pygame so i don't have to rewrite the whole damn thing. the real success here, though, is that i was able to get it to build. that took far longer than it ever should have.